Created linux shell setup scripts.

+++ bin/linux.sh
@@ -2,32 +2,36 @@
 #!/bin/bash
 
 # holding uname as a variable and using -s option to hold actual value
-NAME=$(uname -s)
+NAME=$(uname)
 
 LINUX="Linux"
 # if uname equals Linux and using $ to access their values
-if [ "$NAME" = "$LINUX" ]; then
+if [ "$NAME" != "$LINUX" ]; then
 	# printing a statement saying they are equal
-	echo "uname equals Linux"
-# if uname does not equal linux
-else
-	#error message is redirected to this file
-	2> linuxsetup.log
-	#then the exit command is run
+	echo "uname not equal Linux" >> linuxsetup.log
 	exit
 fi
+# if uname does not equal linux
+#else
+#error message is redirected to this file
+#2> linuxsetup.log
+#then the exit command is run
+#exit
+$fi
 # creating .TRASH directory in home directory if it does not exist
 mkdir -p ~/.TRASH
 
+if [ -f "~/.vimrc" ]; then
 #change the name of .vimrc to .bup_vimrc if the file exists
 mv ~/.vimrc ~/.bup_vimrc
 #printing this statement to the file
 echo "the current .vimrc file ws changed to '.bup_vimrc'" >> linuxsetup.log
+fi
 
 # printing out the contents of vimrc and overwritting it/moving to the .vimrc file
 cat etc/vimrc > ~/.vimrc
 # adding this statement to the end of the .bashrc file
-echo "~/.dotfiles/etc/bashrc_custom" >> ~/.bashrc
+echo "source ~/.dotfiles/etc/bashrc_custom" >> ~/.bashrc
